update: MCPサーバー状態を別ウィジェットへ切り出した
